Abstracted map to its own component as complexity was bloating controller and polished implementation
