+++ client/app/app.scss
@@ -2022,6 +2022,7 @@ input:checked + .slider:before {
 @import '../app/reporting/reporting';
 @import '../app/department-admin/department-admin';
 @import '../app/moveup/moveup-home/moveup-home';
+@import '../components/move-up-map/move-up-map.scss';
 @import '../components/incident/incident';
 @import '../components/oauth-buttons/oauth-buttons';
 @import '../components/percent-change/percent-change';
