+++ client/app/moveup/moveup-home/moveup-home.html
@@ -53,18 +53,22 @@
 
           <div class="model-grid">
             <div class="model-grid-row model-grid-header">
-              <div ng-class="{'txt-black': vm.sort.includes('unit_id') }" ng-click="vm.setSort('unit_id')">Unit<i class="fa fa-angle-down ml-2"></i></div>
-              <div ng-class="{'txt-black': vm.sort.includes('type') }" ng-click="vm.setSort('type')">Unit Type<i class="fa fa-angle-down ml-2"></i></div>
+              <div ng-class="{'txt-black': vm.sort.includes('unit_id') }" ng-click="vm.setSort('unit_id')">Unit<i
+                  class="fa fa-angle-down ml-2"></i></div>
+              <div ng-class="{'txt-black': vm.sort.includes('type') }" ng-click="vm.setSort('type')">Unit Type<i
+                  class="fa fa-angle-down ml-2"></i></div>
               <div>Operational Status</div>
-              <div ng-class="{'txt-black': vm.sort.includes('station') }" ng-click="vm.setSort('station')">Current Location<i class="fa fa-angle-down ml-2"></i></div>
+              <div ng-class="{'txt-black': vm.sort.includes('station') }" ng-click="vm.setSort('station')">Current
+                Location<i class="fa fa-angle-down ml-2"></i></div>
             </div>
 
-            <div class="model-grid-row" ng-repeat="unit in vm.payload.unit_status | orderBy: vm.sort | limitTo : vm.pagination.pageSize : vm.paginationBegin">
+            <div class="model-grid-row"
+              ng-repeat="unit in vm.payload.unit_status | orderBy: vm.sort | limitTo : vm.pagination.pageSize : vm.paginationBegin">
               <div class="flex-start-center">{{unit.unit_id}}</div>
               <div class="flex-start-center">{{unit.type}}</div>
               <div class="flex-start-center">
                 <label class="switch m-0">
-                  <input type="checkbox" ng-model="unit.status">
+                  <input type="checkbox" ng-model="unit.status" ng-change="vm.setDirty()">
                   <span class="slider round red"></span>
                 </label>
                 <span class="ml-2">
@@ -73,23 +77,19 @@
                 </span>
               </div>
               <div>
-                <select class="w180" ng-model="unit.station">
-                  <option ng-repeat="station in vm.stations | orderBy: 'station_number'" value="{{station.station_number}}">Station {{station.station_number}}</option>
+                <select class="w180" ng-model="unit.station" ng-change="vm.setDirty()">
+                  <option ng-repeat="station in vm.stations | orderBy: 'station_number'"
+                    value="{{station.station_number}}">Station {{station.station_number}}</option>
                 </select>
               </div>
             </div>
           </div>
-          
+
           <div class="d-flex justify-content-between p-4">
             <div>
-              <table-controls
-                ng-if="vm.pagination"
-                pagination="vm.pagination"
-                position="bottom"
-                show-sort="false"
-                max-pages="100"
-              ></table-controls>
-            </div> 
+              <table-controls ng-if="vm.pagination" pagination="vm.pagination" position="bottom" show-sort="false"
+                max-pages="100"></table-controls>
+            </div>
             <button class="btn btn-primary" ng-click="vm.run()">Run model</button>
           </div>
         </div>
@@ -99,29 +99,54 @@
     <section class="mt-3">
       <div class="card">
         <div class="card-header">
-          <div ng-class="{ 'model-disabled': vm.outputDisabled }" class="d-flex model-box flex-space-between-center w-100">
+          <div ng-class="{ 'model-disabled': !vm.optimized }" class="d-flex model-box flex-space-between-center w-100">
             <div class="flex-center-center">
               <i class="fa fa-angle-down"></i>
               <h4 class="mb-0 ml-2 mr-2">Model Output</h4>
             </div>
 
-            <div class="optimize-switcher-container" ng-hide="vm.outputDisabled">
+            <div class="optimize-switcher-container" ng-show="vm.optimized">
               Strategy:
               <div class="optimize-switcher">
-                <button ng-click="vm.setStrategy('current')" ng-class="{ 'active': vm.strategy === 'current' }">Current</button>
-                <button ng-click="vm.setStrategy('optimized')" ng-class="{ 'active': vm.strategy === 'optimized' }">Optimized</button>
+                <button ng-click="vm.setStrategy('current')"
+                  ng-class="{ 'active': vm.strategy === 'current' }">Current</button>
+                <button ng-click="vm.setStrategy('optimized')"
+                  ng-class="{ 'active': vm.strategy === 'optimized' }">Optimized</button>
               </div>
             </div>
           </div>
         </div>
+        <div class="card-body p-0 pos-relative" ng-show="vm.optimized">
+          <div ng-if="vm.strategy === 'current'">
+            <move-up-map stations="vm.stations" units="vm.units" />
+            <div class="model-ouput-message flex-center-center flex-column p-4">
+              <div>Currently <span
+                  class="tx-semibold">{{ vm.optimized.current.metrics.percentage_under_4_minute_travel }}%</span> units are under
+                {{ vm.payload.covered_time }} minutes travel time. </div>
+              <a ng-click="vm.setStrategy('optimized')">View the optimized strategy.</a>
+            </div>
+          </div>
 
-        <div class="card-body p-0" ng-hide="vm.outputDisabled">
-          <div id="unoptimized-map">
+          <div ng-if="vm.strategy === 'optimized'">
+            <move-up-map stations="vm.stations" units="vm.units" color="#851bc7"/>
+            <div class="model-ouput-message flex-center-center flex-column p-4">
+              <div>With the optimized strategy, <span class="tx-semibold">{{ vm.optimized.move_up.metrics.percentage_under_4_minute_travel }}%</span> units
+                would be under 4 minutes travel time. Thats an increase of:</div>
+              <div class="model-pill tx-semibold" ng-class="{ red: vm.modelImprovement < 0 }">
+                <i ng-if="vm.modelImprovement >= 0" class="fa fa-arrow-up"></i>
+                <i ng-if="vm.modelImprovement < 0" class="fa fa-arrow-down"></i>
+                {{ vm.modelImprovement }}
+              </div>
+              <div class="tx-semibold">Here are your recommended moves to realize this optimization:</div>
+              <div class="model-move-list">
+                <span ng-repeat="move in vm.optimized.move_up.moves">
+                  - {{ move.unit_id }} moves to Station {{ move.station }}
+                </span>
+              </div>
+            </div>
           </div>
-          <div class="model-ouput-message flex-center-center flex-column p-4">
-            <div>Currently <span class="bold">74.7%</span> units are under {{ vm.payload.covered_time }} minutes travel time. </div>
-            <a>View the optimized strategy.</a>
-          </div>
+
+          <div ng-if="vm.dirty" class="model-change-warning">Run the model in order to view your changes on the map.</div>
         </div>
       </div>
     </section>
