+++ client/app/moveup/moveup-home/moveup-home.scss
@@ -1,5 +1,4 @@
 .moveup-home {
-
   table-controls .root.position-bottom {
     column-gap: 10px;
   }
@@ -95,78 +94,44 @@
       cursor: pointer;
     }
   }
-
-  .bold {
-    font-weight: bold;
-  }
   
-  #unoptimized-map {
-    width: 100%;
-    height: 40vh;
-    position: relative;
-  }
+  .model-pill {
+    border-radius: 24.5px;
+    border: solid 1px rgba(48, 179, 112, 0.16);
+    background-color: #e4f5ec;
+    margin: 20px 0;
+    color: #30b370;
+    background: #e4f5ec;
+    padding: 5px 20px;
 
-  .hide-markers .model-marker {
-    display: none;
-  }
-
-  .layers-button {
-    background: url(/assets/images/icon-layers.svg) no-repeat center;
-    width: 32px;
-    height: 30px;
-    cursor: pointer;
-  }
-
-  .mapboxgl-popup-content {
-    background: rgba(0, 0, 0, 0.87);
-    color: white;
-  }
-
-  .mapboxgl-popup-anchor-top .mapboxgl-popup-tip,
-  .mapboxgl-popup-anchor-bottom .mapboxgl-popup-tip {
-    border-top-color: rgba(0, 0, 0, 0.87);
-    border-bottom-color: rgba(0, 0, 0, 0.87);
-  }
-
-  .model-marker {
-    background: url(/assets/images/fire-station.svg) no-repeat center;
-    width: 32px;
-    height: 40px;
-    cursor: pointer;
-
-    &:after {
-      position: absolute;
-      background: #dc3545;
-      color: white;
-      text-align: center;
-      width: 20px;
-      height: 20px;
-      content: attr(data-count);
-      border-radius: 50%;
-      top: -4px;
-      right: -6px;
+    &.red {
+      color: #fd2a58;
+      background-color: #f5e4e9;
+      border: solid 1px rgba(179, 48, 90, 0.16);
     }
   }
 
-  .model-popup {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-
-    strong {
-      font-size: 18px;
-    }
-
-    .model-popup-unit-count {
-      margin: 10px 0;
-    }
-
-    .model-popup-units {
-      display: flex;
-      width: 100%;
-      justify-content: space-evenly;
-    }
+  .model-move-list {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: auto;
+    grid-gap: 20px;
+    margin-top: 20px;
   }
-  
+
+  .model-change-warning {
+    position: absolute;
+    top: 0;
+    margin-top: 20px;
+    padding: 10px;
+    left: calc(50% - 20%);
+    width: 40%;
+    text-align: center;
+    background: #fff8e8;
+    color: #726c5d;
+    border: 1px solid #e1c786;
+    border-radius: 4px;
+    box-shadow: 0 4px 9px 0 rgba(0, 0, 0, 0.26);
+  }
+
 }
