+++ client/components/move-up-map/move-up-map.html
@@ -1,0 +1,2 @@
+<div id="map">
+</div>
