+++ client/components/move-up-map/move-up-map.scss
@@ -1,0 +1,68 @@
+#map {
+  width: 100%;
+  height: 40vh;
+  position: relative;
+
+  .hide-markers .model-marker {
+    display: none;
+  }
+
+  .layers-button {
+    background: url(/assets/images/icon-layers.svg) no-repeat center;
+    width: 32px;
+    height: 30px;
+    cursor: pointer;
+  }
+
+  .mapboxgl-popup-content {
+    background: rgba(0, 0, 0, 0.87);
+    color: white;
+  }
+
+  .mapboxgl-popup-anchor-top .mapboxgl-popup-tip,
+  .mapboxgl-popup-anchor-bottom .mapboxgl-popup-tip {
+    border-top-color: rgba(0, 0, 0, 0.87);
+    border-bottom-color: rgba(0, 0, 0, 0.87);
+  }
+
+  .model-marker {
+    background: url(/assets/images/fire-station.svg) no-repeat center;
+    width: 32px;
+    height: 40px;
+    cursor: pointer;
+
+    &:after {
+      position: absolute;
+      background: #dc3545;
+      color: white;
+      text-align: center;
+      width: 20px;
+      height: 20px;
+      content: attr(data-count);
+      border-radius: 50%;
+      top: -4px;
+      right: -6px;
+    }
+  }
+
+  .model-popup {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+
+    strong {
+      font-size: 18px;
+    }
+
+    .model-popup-unit-count {
+      margin: 10px 0;
+    }
+
+    .model-popup-units {
+      display: flex;
+      width: 100%;
+      justify-content: space-evenly;
+    }
+  }    
+}
