+++ client/themes/bracket-plus/scss/util/_misc.scss
@@ -24,7 +24,6 @@ $num: 0;
 
 // Flex
 .flex-1 { flex: 1; }
-.flex-column { flex-direction: column; }
 .flex-center {
   display: flex;
   justify-content: center;
